requirements.txt
numpy==1.26.4
pandas==2.3.0
scipy==1.14.1
pyyaml==6.0.2